+++ packages/contract-utils/crypto/src/lib.rs
@@ -6,3 +6,6 @@ pub mod hasher;
 pub mod keccak;
 pub mod merkle;
 pub mod sha256;
+
+#[cfg(test)]
+mod test;
