+++ packages/contract-utils/crypto/src/test/mod.rs
@@ -1,0 +1,4 @@
+mod hashable;
+mod keccak;
+mod merkle;
+mod sha256;
